fix: file structure
