+++ packages/app-runtime/src/natives/events/AppReadyEvent.ts
@@ -1,0 +1,8 @@
+import { NativeEvent } from "../NativeEvent";
+
+export class AppReadyEvent extends NativeEvent {
+    public static namespace = "AppReadyEvent";
+    public constructor() {
+        super(AppReadyEvent.namespace);
+    }
+}
