+++ packages/app-runtime/src/natives/events/UrlOpenEvent.ts
@@ -1,0 +1,8 @@
+import { NativeEvent } from "../NativeEvent";
+
+export class UrlOpenEvent extends NativeEvent {
+    public static namespace = "UrlOpenEvent";
+    public constructor(public readonly url: string) {
+        super(UrlOpenEvent.namespace);
+    }
+}
