+++ packages/app-runtime/src/natives/index.ts
@@ -1,3 +1,4 @@
+export * from "./events";
 export * from "./INativeBootstrapper";
 export * from "./INativeConfigAccess";
 export * from "./INativeDeviceInfoAccess";
@@ -5,4 +6,3 @@ export * from "./INativeEnvironment";
 export * from "./INativeNotificationAccess";
 export * from "./INativeTranslationProvider";
 export * from "./NativeEvent";
-export * from "./NativeEvents";
